+++ src/python/gafam/poster.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
-# (c) 2018-2022 The PosterKit developers <developers@posterkit.org>
+# (c) 2018-2026 The PosterKit developers <developers@posterkit.org>
 import os
 import logging
 import math
-import tempfile
-from io import BytesIO
-from collections import OrderedDict
-from posterkit.makepdf import makepdf
-from posterkit.pdfnup import create_image
-from posterkit.util import ensure_directory, run_command_basic
+from typing import Dict, Any
+
+from posterkit.exposure import convert_image, create_image, pdf_to_svg, layout_pdf, html_to_pdf, concat_pdf_files
+from posterkit.util import ensure_directory, save_buffer
 
 logger = logging.getLogger(__name__)
 
@@ -18,10 +16,12 @@ URI_TEMPLATE = 'https://examples.posterkit.net/lqdn-gafam-campaign/poster.html?l
 #URI_TEMPLATE = 'http://localhost:9999/examples/lqdn-gafam-campaign/poster.html?lang={language}&name={name}&variant={variant}'
 
 # Templates for output files
-PDF_NAME_TEMPLATE       = 'pdf/{variant}/lqdn-gafam-poster-{language}-{variant}.pdf'
-SVG_NAME_TEMPLATE       = 'svg/{variant}/lqdn-gafam-poster-{language}-{variant}/lqdn-gafam-poster-{language}-{variant}-{number:0>2}-{name}.svg'
-THUMBNAIL_NAME_TEMPLATE = 'img/{variant}/lqdn-gafam-poster-{language}-{variant}-{nup}-{size}.{suffix}'
-MOSAIC_NAME_TEMPLATE    = 'img/mosaic/lqdn-gafam-poster-mosaic-{variant}-{direction}.{suffix}'
+PROJECT_NAME            = 'lqdn-gafam-poster'
+SINGLE_PAGE_TEMPLATE    = 'page/{suffix}/{variant}/{project}-{language}-{variant}/{project}-{language}-{variant}-{number:0>2}-{name}.{suffix}'
+SERIES_PDF_TEMPLATE     = 'series/{variant}/{project}-{language}-{variant}.pdf'
+SERIES_SVG_TEMPLATE     = 'series/{variant}/{project}-{language}-{variant}-{nup}.{suffix}'
+SERIES_BITMAP_TEMPLATE  = 'series/{variant}/{project}-{language}-{variant}-{nup}-{size}.{suffix}'
+MOSAIC_NAME_TEMPLATE    = 'mosaic/{project}-mosaic-{variant}-{direction}.{suffix}'
 
 # Which poster names are "one set"
 POSTER_NAMES = [
@@ -41,142 +41,129 @@ POSTER_VARIANTS = [
 # Path to GitHub API for listing the directory contents where the i18n json files are stored
 POSTER_TRANSLATIONS_URI = 'https://api.github.com/repos/gafam/gafam-poster-translations/contents/json'
 
-# Use False for debugging to keep all temporary files
-DELETE_TEMPFILES = True
-
 
 def render_posters(info=None, path=None):
+    """Render both single-page and composite documents"""
 
-    logger.info('Rendering GAFAM campaign posters')
+    logger.info('Rendering campaign posters')
     logger.info('The options are {}'.format(info))
     logger.info('The output path is {}'.format(path))
 
-    # TODO: What about reasonable defaults?
+    # TODO: What about reasonable shortcut defaults?
     #variant = variant or 'default'
 
-    # Aggregate all PDF files
-    pdf_files = []
+    render_pages(info, path)
+    render_composites(info, path)
+
+
+def render_pages(info, path):
+    """Generate multiple single-page PDF files and SVG documents."""
+
+    context: Dict[str, Any] = {
+        "project": PROJECT_NAME,
+    }
+
+    for language in info['language']:
+        logger.info(f'* Rendering single-page PDF posters for language={language}')
+        for variant in info['variant']:
+            logger.info(f'** Rendering single-page PDF posters for variant={variant}')
+            for index, name in enumerate(info['name']):
+                number = index + 1
+                logger.info(f'*** Rendering single-page PDF poster with language={language}, variant={variant}, name={name}')
+
+                context.update({"language": language, "variant": variant, "name": name, "number": number})
+
+                # Compute resource URL
+                uri = URI_TEMPLATE.format(**context)
+
+                # Single-page PDF
+                pdf_single_filename = SINGLE_PAGE_TEMPLATE.format(**context, suffix="pdf")
+                pdf_single_path = os.path.join(path, pdf_single_filename)
+                logging.info('PDF path: {}'.format(pdf_single_path))
+                if not (os.path.exists(pdf_single_path) and os.path.getsize(pdf_single_path) > 0):
+                    stream = html_to_pdf(uri)
+                    save_buffer(stream, pdf_single_path)
+
+                # Single-page SVG
+                svg_filename = SINGLE_PAGE_TEMPLATE.format(**context, suffix="svg")
+                svg_filepath = os.path.abspath(os.path.join(path, svg_filename))
+                logging.info('SVG path: {}'.format(svg_filepath))
+                if not (os.path.exists(svg_filepath) and os.path.getsize(svg_filepath) > 0):
+                    # Ensure path exists
+                    ensure_directory(svg_filepath)
+                    # Convert PDF to SVG
+                    pdf_to_svg(pdf=pdf_single_path, outputfile=svg_filepath)
+
+                # Single-page PNG
+                png_filename = SINGLE_PAGE_TEMPLATE.format(**context, suffix="png")
+                png_filepath = os.path.abspath(os.path.join(path, png_filename))
+                stream = convert_image(input_file=pdf_single_path, format="png")
+                save_buffer(stream, png_filepath)
+
+
+def render_composites(info=None, path=None):
+    """Generate composite PDF files and images."""
+
+    context = {
+        "project": PROJECT_NAME,
+    }
+
+    nup_options = {
+        "papersize": '297mm,1050mm',
+        "nup": "5x1",
+    }
+
+    context.update(nup_options)
 
     # Generate multi-page PDF files
     for language in info['language']:
-
-        logger.info('* Rendering PDF posters for language={language}'.format(**locals()))
-
+        logger.info(f'* Rendering PDF poster series for language={language}')
         for variant in info['variant']:
+            logger.info(f'** Rendering PDF poster series for variant={variant}')
 
-            logger.info('** Rendering PDF posters for variant={variant}'.format(**locals()))
+            context.update({"language": language, "variant": variant})
 
             # Compute output path
-            if path != '-':
-                pdf_filename = PDF_NAME_TEMPLATE.format(**locals())
-                pdf_filepath = os.path.abspath(os.path.join(path, pdf_filename))
+            pdf_filename = SERIES_PDF_TEMPLATE.format(**context)
+            pdf_filepath = os.path.abspath(os.path.join(path, pdf_filename))
+            logging.info('PDF path: {}'.format(pdf_filepath))
 
-            if os.path.exists(pdf_filepath):
-                logger.info(f"{pdf_filepath} already exists, skipping all further rendering steps")
-                continue
+            # A. Composite PDF (regular)
+            pdf_files = []
+            for index, name in enumerate(info['name']):
+                pdf_single_filename = SINGLE_PAGE_TEMPLATE.format(**context, name=name, number=index + 1, suffix="pdf")
+                pdf_single_path = os.path.join(path, pdf_single_filename)
+                pdf_files.append(pdf_single_path)
+            stream = concat_pdf_files(pdf_files)
+            save_buffer(stream, pdf_filepath)
 
-            # Generate single-page PDF files
-            # Render N pages
-            outputfiles = []
-            outputmap = OrderedDict()
-            for name in info['name']:
+            # B. Composite SVG (wide)
+            pdf_nup = layout_pdf(pdf_files, **nup_options)
+            svg_filename = SERIES_SVG_TEMPLATE.format(**context, suffix="svg")
+            svg_filepath = os.path.abspath(os.path.join(path, svg_filename))
+            pdf_to_svg(pdf=pdf_nup, outputfile=svg_filepath)
 
-                logger.info('*** Rendering PDF poster with language={language}, variant={variant}, name={name}'.format(**locals()))
-
-                # Compute resource URL
-                uri = URI_TEMPLATE.format(
-                    language=language,
-                    name=name,
-                    variant=variant)
-
-                tmpfile = tempfile.NamedTemporaryFile(delete=DELETE_TEMPFILES)
-
-                # RENDER RENDER RENDER
-
-                # Render PDF
-                stream = makepdf(uri)
-
-                # Write to temporary file
-                tmpfile.write(stream.read())
-                tmpfile.flush()
-
-                # Bookkeeping
-                outputfiles.append(tmpfile)
-                outputmap[name] = tmpfile
-
-            # Join N pages into single document
-            outputfilenames = [item.name for item in outputfiles]
-            pdfstream = join_pdf_files(outputfilenames)
-
-            # Write to STDOUT
-            if path == '-':
-                print(pdfstream.read())
-
-            # Save to filesystem
-            else:
-
-                # A. Save PDF
-                logging.info('PDF file path is {}'.format(pdf_filepath))
-
-                # Ensure path exists
-                ensure_directory(pdf_filepath)
-
-                # Save PDF
-                with open(pdf_filepath, 'wb') as f:
-                    f.write(pdfstream.read())
-
-                # Remember PDF path
-                pdf_files.append(pdf_filepath)
-
-
-                # B. Create thumbnail images
-                format = 'png8'
-                suffix = 'png'
-                nup = '5x1'
-                #sizes = ['640x', '1280x', '1920x', '2560x']
-                sizes = ['1280x', '2560x']
-
-                for size in sizes:
-
-                    # Render PNG image from multi-page PDF document
-                    image = create_image(pdf_filepath, papersize='297mm,1050mm', nup=nup, size=size, format=format)
-
-                    # Compute filename
-                    img_filename = THUMBNAIL_NAME_TEMPLATE.format(**locals())
-                    img_filepath = os.path.abspath(os.path.join(path, img_filename))
-
-                    # Save image
-                    filepath = save_file(image, path, img_filename)
-                    logger.info('Saved thumbnail image to {}'.format(filepath))
-
-
-                # C. Convert to single-page SVG image
-
-                number = 1
-                for name, pdfpage_file in outputmap.items():
-
-                    # Compute output path
-                    svg_filename = SVG_NAME_TEMPLATE.format(**locals())
-                    svg_filepath = os.path.abspath(os.path.join(path, svg_filename))
-                    logging.info('SVG file path is {}'.format(svg_filepath))
-
-                    # Ensure path exists
-                    ensure_directory(svg_filepath)
-
-                    # Convert PDF to SVG
-                    command = "pdf2svg '{inputfile}' '{outputfile}'".format(inputfile=pdfpage_file.name, outputfile=svg_filepath)
-                    logger.info(u'The PDF to SVG conversion command is: {}'.format(command))
-                    if not run_command_basic(command):
-                        logger.warning("PDF to SVG conversion failed")
-
-                    # Continue with next page
-                    number += 1
-
-
-    return pdf_files
+            # C. Composite bitmap images (wide)
+            bitmap_variant = {
+                #"sizes": ['640x', '1280x', '1920x', '2560x'],
+                "sizes": ['1280x', '2560x']
+            }
+            for size in bitmap_variant["sizes"]:
+                image = create_image(pdf_filepath, size=size, **nup_options)
+                img_filename = SERIES_BITMAP_TEMPLATE.format(**context, size=size, suffix="png")
+                img_filepath = os.path.abspath(os.path.join(path, img_filename))
+                logging.info('Image file path: {}'.format(img_filepath))
+                #if os.path.exists(img_filepath) and os.path.getsize(img_filepath) > 0:
+                #    continue
+                save_buffer(image, img_filepath)
 
 
 def render_mosaic(path=None, files=None, variant=None):
+
+    context = {
+        "project": PROJECT_NAME,
+        "variant": variant,
+    }
 
     logger.info('Rendering mosaic image')
     #logger.info('The options are {}'.format(info))
@@ -214,43 +201,19 @@ def render_mosaic(path=None, files=None, variant=None):
 
 
     # Render PNG image from multiple multi-page PDF documents
-    format = 'png8'
+    format = 'png'
     image = create_image(pdf_selected, papersize=papersize, nup=nup, size=None, format=format)
 
     # Compute filename
-    direction = f'{number_of_columns}-columns'
-    suffix = 'png'
-    img_filename = MOSAIC_NAME_TEMPLATE.format(**locals())
+    context.update({
+        "direction": f'{number_of_columns}-columns',
+        "suffix": 'png',
+    })
+
+    img_filename = MOSAIC_NAME_TEMPLATE.format(**context)
 
     # Save image
     # TODO: Optionally optimize image using `oxipng`.
-    filepath = save_file(image, path, img_filename)
-    logger.info('Saved mosaic image to {}'.format(filepath))
-
-
-def save_file(buffer, path, filename):
-
-    # Compute full path
-    img_filepath = os.path.abspath(os.path.join(path, filename))
-
-    # Ensure path exists
-    ensure_directory(img_filepath)
-
-    # Save to filesystem
-    with open(img_filepath, 'wb') as f:
-        f.write(buffer.read())
-
-    return img_filepath
-
-
-def join_pdf_files(filenames):
-    input_files = ' '.join(filenames)
-    tmpfile = tempfile.NamedTemporaryFile(delete=DELETE_TEMPFILES)
-    output_file = tmpfile.name
-    join_command = 'pdftk {input_files} output {output_file}'.format(**locals())
-    logger.info(u'The joining command is: {}'.format(join_command))
-    if not run_command_basic(join_command):
-        logger.warning("Joining PDF files failed")
-    tmpfile.seek(0)
-    buffer = BytesIO(tmpfile.read())
-    return buffer
+    img_filepath = os.path.abspath(os.path.join(path, img_filename))
+    logger.info('Mosaic image path: {}'.format(img_filepath))
+    save_buffer(image, img_filepath)