+++ src/python/posterkit/commands.py
@@ -1,10 +1,14 @@
 # -*- coding: utf-8 -*-
 # (c) 2018 The PosterKit developers <developers@posterkit.org>
 import logging
+import sys
+from tempfile import NamedTemporaryFile
+
+import requests
 from docopt import docopt
 from posterkit import __version__
-from posterkit.makepdf import makepdf
-from posterkit.pdfnup import create_image
+from posterkit.exposure import create_image, html_to_pdf
+from posterkit.settings import DELETE_TEMPFILES
 from posterkit.util import boot_logging, normalize_options
 
 logger = logging.getLogger(__name__)
@@ -15,7 +19,7 @@ def run():
     """
     Usage:
         posterkit pdf --url=<url>
-        posterkit thumbnail --pdf=<pdf> --nup=1x5 [--size=640x] [--format=png]
+        posterkit thumbnail --pdf=<pdf> [--nup=1x5] [--size=640x] [--format=png]
         posterkit --help
 
     Options for "pdf":
@@ -23,7 +27,7 @@ def run():
 
     Options for "thumbnail":
         --pdf=<pdf>             Input PDF file to use for generating thumbnail image.
-        --nup=<nup>             N-up directory how to layout the thumbnail. [default: 1]
+        --nup=<nup>             N-up directive how to lay out the thumbnail. [default: 1]
         --size=<size>           Size of the generated image. [default: 1024x]
         --format=<format>       Output image format. [default: jpg]
 
@@ -32,6 +36,8 @@ def run():
         # Render single-page PDF document and output to STDOUT
         posterkit pdf --url=https://examples.posterkit.net/rfa-endlich-normale-menschen/poster.html
 
+        # Render PDF
+        posterkit thumbnail --pdf=lqdn-gafam-poster-de-color-01-google.pdf
     """
 
     # Use generic commandline options schema and amend with current program name
@@ -43,16 +49,20 @@ def run():
     # Initialize logging
     boot_logging(options)
 
-    # Normalize commandline options
-    options = normalize_options(options)
-
     # Render PDF
     if options['pdf']:
-        stream = makepdf(options['url'])
+        stream = html_to_pdf(options['--url'])
 
         # Write to STDOUT
-        print(stream.read())
+        sys.stdout.buffer.write(stream.getvalue())
 
+    # Render bitmap
     elif options['thumbnail']:
-        image = create_image(options['pdf'], nup=options['nup'], size=options['size'], format=options['format'])
-        print(image.read())
+        options = normalize_options(options)
+        pdf = options['pdf']
+        tmpfile = NamedTemporaryFile(suffix=".pdf", delete=DELETE_TEMPFILES)
+        if pdf.startswith('http'):
+            tmpfile.write(requests.get(pdf).content)
+            pdf = tmpfile.name
+        image = create_image(pdf, nup=options['nup'], size=options['size'], format=options['format'])
+        sys.stdout.buffer.write(image.getvalue())